Split Colonini's AssignStatement to new file
